Update directory structure, moved user schema to user.js, added app.js and edited server\server.js and models\user.js
